pyproject.toml
[project]
name = "pod"
version = "0.1.0"
description = ""
authors = [
    {name = "Phong Nguyen Duong",email = "pnguyenduong@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-py (>=2.6.4,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "aiosqlite (>=0.22.0,<0.23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:pytest_asyncio.*",
]
